requirements.txt
numpy~=2.0.0
pillow~=10.4.0
tqdm~=4.66.4